+++ cpp/libs/engine/core/message_limiter/message_limiter.h
@@ -7,13 +7,13 @@ class MessageLimiter {
 public:
     MessageLimiter() {}
 
-    MessageLimiter(uint8_t defaultMaxRequests, uint64_t defaultTimeWindow);
+    MessageLimiter(uint8_t defaultMaxRequests, uint32_t defaultTimeWindow);
 
     uint32_t CanSend(uint32_t messageId);
     
 private:
     uint8_t defaultMaxRequests{3};  // max requests per window
-    uint64_t defaultTimeWindow{1};   // window duration (seconds)
+    uint32_t defaultTimeWindow{1};   // window duration (seconds)
     std::unordered_map<uint32_t, boost::circular_buffer<uint64_t>> requestRecords;
 };
